+++ src/lib/src/core/db/data_frames.rs
@@ -1,2 +1,4 @@
 pub mod df_db;
+pub mod rows;
+pub mod columns;
 pub mod workspace_df_db;
