+++ src/lib/src/view/data_frames.rs
@@ -1,6 +1,15 @@
 use serde::{Deserialize, Serialize};
 
+pub mod columns;
+
 #[derive(Deserialize, Serialize, Debug)]
 pub struct DataFramePayload {
     pub is_indexed: bool,
+}
+
+#[derive(Deserialize, Serialize, Debug)]
+pub struct DataFrameColumnChange {
+    pub column_name: String,
+    pub operation: String,
+    pub new_name: String,
 }
