+++ src/lib/src/view/data_frames/columns.rs
@@ -1,0 +1,7 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Deserialize, Serialize, Debug)]
+pub struct NewColumn {
+    pub name: String,
+    pub data_type: String,
+}
