+++ src/server/src/errors.rs
@@ -324,6 +324,21 @@ impl error::ResponseError for OxenHttpError {
                         });
                         HttpResponse::BadRequest().json(error_json)
                     }
+                    OxenError::ColumnNameAlreadyExists(column_name) => {
+                        log::error!("Column Name Already Exists schemas: {}", column_name);
+                        let error_json = json!({
+                            "error": {
+                                "type": "column_error",
+                                "title":
+                                    "Column Name Already Exists",
+                                "detail":
+                                    format!("Column name '{}' already exists in schema", column_name)
+                            },
+                            "status": STATUS_ERROR,
+                            "status_message": MSG_BAD_REQUEST,
+                        });
+                        HttpResponse::BadRequest().json(error_json)
+                    }
                     OxenError::DUCKDB(error) => {
                         log::error!("DuckDB error: {}", error);
